refactor(scim): split groups.go into per-RPC files (#149 slice 5)
Fifth slice. Mirrors slice 4's users.go split applied to groups.go.

- groups_create.go: createGroup (the SCIM-client-friendly idempotent
  POST handler that syncs display name + members on re-POST instead
  of returning 409)
- groups_mutate.go: replaceGroup, patchGroup,
  handleGroupPatch{Add,Remove,Replace} (PUT + PATCH dispatchers)
- groups.go: listGroups, listGroupsFiltered, getGroup, deleteGroup only

Size delta:
- groups.go: 780 \xe2\x86\x92 241 LOC (-539) \xe2\x86\x90 ACCEPTANCE BAR REACHED
- New: groups_create.go (219), groups_mutate.go (358)
- Cumulative since baseline 1006 LOC: -765 (-76%)

Pure file-shuffling \xe2\x80\x94 zero behaviour change. Stacked on #233 (slice
3 groups_helpers); will rebase as it lands.

Refs manchtools/power-manage-server#149 (audit F009).
