+++ internal/scim/groups.go
@@ -1,13 +1,10 @@
 package scim
 
 import (
-	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
 	"strconv"
-	"strings"
 
 	"github.com/jackc/pgx/v5"
 
@@ -142,204 +139,6 @@ func (h *Handler) listGroupsFiltered(w http.ResponseWriter, r *http.Request, pro
 }
 
 // createGroup handles POST /scim/v2/{slug}/Groups
-func (h *Handler) createGroup(w http.ResponseWriter, r *http.Request) {
-	h.logger.Debug("SCIM createGroup called")
-	provider, ok := providerFromContext(r.Context())
-	if !ok {
-		writeError(w, http.StatusUnauthorized, "not authenticated")
-		return
-	}
-
-	var scimGroup SCIMGroup
-	limitBody(r)
-	if err := json.NewDecoder(r.Body).Decode(&scimGroup); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-
-	if scimGroup.DisplayName == "" {
-		writeError(w, http.StatusBadRequest, "displayName is required")
-		return
-	}
-
-	ctx := r.Context()
-	baseURL := baseURLFromRequest(r, provider.Slug)
-
-	scimGroupID := scimGroup.ExternalID
-	if scimGroupID == "" {
-		// Use the SCIM-generated ID as the SCIM group ID if no external ID
-		scimGroupID = scimGroup.ID
-	}
-	if scimGroupID == "" {
-		scimGroupID = newULID()
-	}
-
-	// Check if this SCIM group is already mapped
-	existing, err := h.store.Queries().GetSCIMGroupMapping(ctx, db.GetSCIMGroupMappingParams{
-		ProviderID:  provider.ID,
-		ScimGroupID: scimGroupID,
-	})
-	if err == nil {
-		// Already exists — update display name if changed and return existing resource.
-		h.logger.Debug("SCIM createGroup: group already exists, syncing", "scim_group_id", scimGroupID, "user_group_id", existing.UserGroupID)
-		// This makes POST idempotent, which handles SCIM clients that re-POST on every sync.
-		if existing.ScimDisplayName != scimGroup.DisplayName {
-			h.appendEvent(ctx, store.Event{
-				StreamType: "scim_group_mapping",
-				StreamID:   existing.ID,
-				EventType:  string(eventtypes.SCIMGroupMappingUpdated),
-				Data: map[string]any{
-					"provider_id":       provider.ID,
-					"scim_group_id":     scimGroupID,
-					"scim_display_name": scimGroup.DisplayName,
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-			h.appendEvent(ctx, store.Event{
-				StreamType: "user_group",
-				StreamID:   existing.UserGroupID,
-				EventType:  string(eventtypes.UserGroupUpdated),
-				Data: map[string]any{
-					"name": scimGroup.DisplayName,
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-		}
-
-		// Reconcile members if the field was present in the JSON body.
-		// nil means the field was omitted (don't touch members).
-		// Empty slice means explicitly empty (remove all members).
-		if scimGroup.Members != nil {
-			h.reconcileGroupMembers(ctx, provider, existing.UserGroupID, scimGroup.Members)
-		}
-
-		group, err := h.buildGroupResource(ctx, provider.ID, existing, baseURL)
-		if err != nil {
-			// User group referenced by mapping no longer exists.
-			// Remove the orphaned mapping and fall through to create
-			// a fresh group+mapping pair below.
-			h.logger.Warn("orphaned SCIM group mapping, removing and recreating",
-				"mapping_id", existing.ID, "user_group_id", existing.UserGroupID, "error", err)
-			h.appendEvent(ctx, store.Event{
-				StreamType: "scim_group_mapping",
-				StreamID:   existing.ID,
-				EventType:  string(eventtypes.SCIMGroupUnmapped),
-				Data: map[string]any{
-					"provider_id":   provider.ID,
-					"scim_group_id": existing.ScimGroupID,
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-			// Fall through to create a new group below
-		} else {
-			writeJSON(w, http.StatusOK, group)
-			return
-		}
-	}
-	if !errors.Is(err, pgx.ErrNoRows) && err != nil {
-		h.logger.Error("failed to check existing SCIM group mapping", "error", err)
-		writeError(w, http.StatusInternalServerError, "internal server error")
-		return
-	}
-
-	// Create the user group
-	h.logger.Debug("SCIM createGroup: creating new group", "scim_group_id", scimGroupID, "display_name", scimGroup.DisplayName)
-	userGroupID := newULID()
-	err = h.store.AppendEvent(ctx, store.Event{
-		StreamType: "user_group",
-		StreamID:   userGroupID,
-		EventType:  string(eventtypes.UserGroupCreated),
-		Data: map[string]any{
-			"name":        scimGroup.DisplayName,
-			"description": fmt.Sprintf("SCIM-provisioned group from %s", provider.Name),
-		},
-		ActorType: "scim",
-		ActorID:   provider.ID,
-	})
-	if err != nil {
-		h.logger.Error("failed to create user group via SCIM", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to create group")
-		return
-	}
-
-	// Create the SCIM group mapping
-	mappingID := newULID()
-	err = h.store.AppendEvent(ctx, store.Event{
-		StreamType: "scim_group_mapping",
-		StreamID:   mappingID,
-		EventType:  string(eventtypes.SCIMGroupMapped),
-		Data: map[string]any{
-			"provider_id":       provider.ID,
-			"scim_group_id":     scimGroupID,
-			"scim_display_name": scimGroup.DisplayName,
-			"user_group_id":     userGroupID,
-		},
-		ActorType: "scim",
-		ActorID:   provider.ID,
-	})
-	if err != nil {
-		h.logger.Error("failed to create SCIM group mapping", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to map SCIM group")
-		return
-	}
-
-	// Add members if provided
-	for _, member := range scimGroup.Members {
-		if member.Value == "" {
-			continue
-		}
-		streamID := userGroupID + ":" + member.Value
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user_group",
-			StreamID:   streamID,
-			EventType:  string(eventtypes.UserGroupMemberAdded),
-			Data: map[string]any{
-				"group_id": userGroupID,
-				"user_id":  member.Value,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-	}
-
-	// Build response
-	mapping, err := h.store.Queries().GetSCIMGroupMapping(ctx, db.GetSCIMGroupMappingParams{
-		ProviderID:  provider.ID,
-		ScimGroupID: scimGroupID,
-	})
-	if err != nil {
-		// Fall back to a minimal response
-		writeJSON(w, http.StatusCreated, SCIMGroup{
-			Schemas:     []string{GroupSchema},
-			ID:          userGroupID,
-			ExternalID:  scimGroupID,
-			DisplayName: scimGroup.DisplayName,
-			Meta: &SCIMMeta{
-				ResourceType: "Group",
-				Location:     baseURL + "/Groups/" + userGroupID,
-			},
-		})
-		return
-	}
-
-	group, err := h.buildGroupResource(ctx, provider.ID, mapping, baseURL)
-	if err != nil {
-		h.logger.Error("failed to build group resource after create", "error", err)
-		writeJSON(w, http.StatusCreated, SCIMGroup{
-			Schemas:     []string{GroupSchema},
-			ID:          userGroupID,
-			ExternalID:  scimGroupID,
-			DisplayName: scimGroup.DisplayName,
-		})
-		return
-	}
-
-	writeJSON(w, http.StatusCreated, group)
-}
-
 // getGroup handles GET /scim/v2/{slug}/Groups/{id}
 func (h *Handler) getGroup(w http.ResponseWriter, r *http.Request) {
 	h.logger.Debug("SCIM getGroup called")
@@ -383,344 +182,6 @@ func (h *Handler) getGroup(w http.ResponseWriter, r *http.Request) {
 }
 
 // replaceGroup handles PUT /scim/v2/{slug}/Groups/{id}
-func (h *Handler) replaceGroup(w http.ResponseWriter, r *http.Request) {
-	h.logger.Debug("SCIM replaceGroup called")
-	provider, ok := providerFromContext(r.Context())
-	if !ok {
-		writeError(w, http.StatusUnauthorized, "not authenticated")
-		return
-	}
-
-	groupID := r.PathValue("id")
-	if groupID == "" {
-		writeError(w, http.StatusBadRequest, "missing group id")
-		return
-	}
-
-	var scimGroup SCIMGroup
-	limitBody(r)
-	if err := json.NewDecoder(r.Body).Decode(&scimGroup); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-
-	ctx := r.Context()
-	baseURL := baseURLFromRequest(r, provider.Slug)
-
-	// Look up the SCIM group mapping
-	mapping, err := h.store.Queries().GetSCIMGroupMappingByUserGroup(ctx, db.GetSCIMGroupMappingByUserGroupParams{
-		ProviderID:  provider.ID,
-		UserGroupID: groupID,
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			writeError(w, http.StatusNotFound, "group not found")
-			return
-		}
-		h.logger.Error("failed to get SCIM group mapping for replace", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to get group")
-		return
-	}
-
-	// Update display name if changed
-	if scimGroup.DisplayName != "" && scimGroup.DisplayName != mapping.ScimDisplayName {
-		h.appendEvent(ctx, store.Event{
-			StreamType: "scim_group_mapping",
-			StreamID:   mapping.ID,
-			EventType:  string(eventtypes.SCIMGroupMappingUpdated),
-			Data: map[string]any{
-				"provider_id":       provider.ID,
-				"scim_group_id":     mapping.ScimGroupID,
-				"scim_display_name": scimGroup.DisplayName,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user_group",
-			StreamID:   groupID,
-			EventType:  string(eventtypes.UserGroupUpdated),
-			Data: map[string]any{
-				"name": scimGroup.DisplayName,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-	}
-
-	// Reconcile members if the field was present in the JSON body.
-	// nil means the field was omitted (don't touch members).
-	// Empty slice means explicitly empty (remove all members).
-	// SCIM is treated as the source of truth for group membership.
-	if scimGroup.Members != nil {
-		h.reconcileGroupMembers(ctx, provider, groupID, scimGroup.Members)
-	}
-
-	// Read back and return
-	updatedMapping, err := h.store.Queries().GetSCIMGroupMappingByUserGroup(ctx, db.GetSCIMGroupMappingByUserGroupParams{
-		ProviderID:  provider.ID,
-		UserGroupID: groupID,
-	})
-	if err != nil {
-		updatedMapping = mapping
-	}
-
-	group, err := h.buildGroupResource(ctx, provider.ID, updatedMapping, baseURL)
-	if err != nil {
-		h.logger.Error("failed to build group resource after replace", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to build group")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, group)
-}
-
-// patchGroup handles PATCH /scim/v2/{slug}/Groups/{id}
-func (h *Handler) patchGroup(w http.ResponseWriter, r *http.Request) {
-	h.logger.Debug("SCIM patchGroup called")
-	provider, ok := providerFromContext(r.Context())
-	if !ok {
-		writeError(w, http.StatusUnauthorized, "not authenticated")
-		return
-	}
-
-	groupID := r.PathValue("id")
-	if groupID == "" {
-		writeError(w, http.StatusBadRequest, "missing group id")
-		return
-	}
-
-	limitBody(r)
-	var patch SCIMPatchRequest
-	if err := json.NewDecoder(r.Body).Decode(&patch); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid JSON body")
-		return
-	}
-
-	ctx := r.Context()
-	baseURL := baseURLFromRequest(r, provider.Slug)
-
-	// Verify group exists
-	mapping, err := h.store.Queries().GetSCIMGroupMappingByUserGroup(ctx, db.GetSCIMGroupMappingByUserGroupParams{
-		ProviderID:  provider.ID,
-		UserGroupID: groupID,
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			writeError(w, http.StatusNotFound, "group not found")
-			return
-		}
-		h.logger.Error("failed to get SCIM group mapping for patch", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to get group")
-		return
-	}
-
-	for _, op := range patch.Operations {
-		// Pre-validate per RFC 7644 §3.5.2 — `op` must be one of
-		// add | remove | replace. Reject up front so unknown verbs
-		// never silently fall through to a per-op switch default
-		// elsewhere in the codebase.
-		if !op.Op.IsValid() {
-			writeError(w, http.StatusBadRequest, fmt.Sprintf("unsupported patch op: %s", op.Op))
-			return
-		}
-		switch op.Op.Normalize() {
-		case SCIMPatchOpAdd:
-			h.handleGroupPatchAdd(ctx, provider, groupID, op)
-		case SCIMPatchOpRemove:
-			h.handleGroupPatchRemove(ctx, provider, groupID, op)
-		case SCIMPatchOpReplace:
-			h.handleGroupPatchReplace(ctx, provider, groupID, mapping, op)
-		default:
-			// Unreachable: IsValid above guards this switch.
-			writeError(w, http.StatusBadRequest, fmt.Sprintf("unsupported patch op: %s", op.Op))
-			return
-		}
-	}
-
-	// Read back and return
-	updatedMapping, err := h.store.Queries().GetSCIMGroupMappingByUserGroup(ctx, db.GetSCIMGroupMappingByUserGroupParams{
-		ProviderID:  provider.ID,
-		UserGroupID: groupID,
-	})
-	if err != nil {
-		updatedMapping = mapping
-	}
-
-	group, err := h.buildGroupResource(ctx, provider.ID, updatedMapping, baseURL)
-	if err != nil {
-		h.logger.Error("failed to build group resource after patch", "error", err)
-		writeError(w, http.StatusInternalServerError, "failed to build group")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, group)
-}
-
-// handleGroupPatchAdd processes an "add" patch operation on a group.
-func (h *Handler) handleGroupPatchAdd(ctx context.Context, provider db.IdentityProvidersProjection, groupID string, op SCIMPatchOp) {
-	path := strings.ToLower(op.Path)
-	if path != "members" && path != "" {
-		return
-	}
-
-	members := extractMembers(op.Value)
-	for _, userID := range members {
-		streamID := groupID + ":" + userID
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user_group",
-			StreamID:   streamID,
-			EventType:  string(eventtypes.UserGroupMemberAdded),
-			Data: map[string]any{
-				"group_id": groupID,
-				"user_id":  userID,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-	}
-}
-
-// handleGroupPatchRemove processes a "remove" patch operation on a group.
-func (h *Handler) handleGroupPatchRemove(ctx context.Context, provider db.IdentityProvidersProjection, groupID string, op SCIMPatchOp) {
-	path := strings.ToLower(op.Path)
-
-	// Handle path like: members[value eq "userId"]
-	if strings.HasPrefix(path, "members[") {
-		// Extract user ID from the ORIGINAL path to preserve case
-		userID := extractUserIDFromMemberFilter(op.Path)
-		if userID != "" {
-			streamID := groupID + ":" + userID
-			h.appendEvent(ctx, store.Event{
-				StreamType: "user_group",
-				StreamID:   streamID,
-				EventType:  string(eventtypes.UserGroupMemberRemoved),
-				Data: map[string]any{
-					"group_id": groupID,
-					"user_id":  userID,
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-		}
-		return
-	}
-
-	// Handle path "members" with value containing member list
-	if path == "members" || path == "" {
-		members := extractMembers(op.Value)
-		for _, userID := range members {
-			streamID := groupID + ":" + userID
-			h.appendEvent(ctx, store.Event{
-				StreamType: "user_group",
-				StreamID:   streamID,
-				EventType:  string(eventtypes.UserGroupMemberRemoved),
-				Data: map[string]any{
-					"group_id": groupID,
-					"user_id":  userID,
-				},
-				ActorType: "scim",
-				ActorID:   provider.ID,
-			})
-		}
-	}
-}
-
-// handleGroupPatchReplace processes a "replace" patch operation on a group.
-func (h *Handler) handleGroupPatchReplace(ctx context.Context, provider db.IdentityProvidersProjection, groupID string, mapping db.ScimGroupMappingProjection, op SCIMPatchOp) {
-	path := strings.ToLower(op.Path)
-
-	switch path {
-	case "displayname":
-		name, ok := op.Value.(string)
-		if !ok || name == "" {
-			return
-		}
-
-		h.appendEvent(ctx, store.Event{
-			StreamType: "scim_group_mapping",
-			StreamID:   mapping.ID,
-			EventType:  string(eventtypes.SCIMGroupMappingUpdated),
-			Data: map[string]any{
-				"provider_id":       provider.ID,
-				"scim_group_id":     mapping.ScimGroupID,
-				"scim_display_name": name,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user_group",
-			StreamID:   groupID,
-			EventType:  string(eventtypes.UserGroupUpdated),
-			Data: map[string]any{
-				"name": name,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-
-	case "members":
-		// Full member replacement
-		members := extractMembers(op.Value)
-		requestedSet := make(map[string]bool, len(members))
-		for _, id := range members {
-			requestedSet[id] = true
-		}
-
-		currentMemberIDs, err := h.store.Queries().ListUserGroupMemberIDs(ctx, groupID)
-		if err != nil {
-			h.logger.Error("failed to list group members for patch replace", "group_id", groupID, "error", err)
-			return
-		}
-
-		currentSet := make(map[string]bool, len(currentMemberIDs))
-		for _, id := range currentMemberIDs {
-			currentSet[id] = true
-		}
-
-		// Add new members
-		for _, userID := range members {
-			if !currentSet[userID] {
-				h.logger.Debug("SCIM adding member to group", "group_id", groupID, "user_id", userID)
-				streamID := groupID + ":" + userID
-				h.appendEvent(ctx, store.Event{
-					StreamType: "user_group",
-					StreamID:   streamID,
-					EventType:  string(eventtypes.UserGroupMemberAdded),
-					Data: map[string]any{
-						"group_id": groupID,
-						"user_id":  userID,
-					},
-					ActorType: "scim",
-					ActorID:   provider.ID,
-				})
-			}
-		}
-
-		// Remove old members
-		for _, userID := range currentMemberIDs {
-			if !requestedSet[userID] {
-				h.logger.Debug("SCIM removing member from group", "group_id", groupID, "user_id", userID)
-				streamID := groupID + ":" + userID
-				h.appendEvent(ctx, store.Event{
-					StreamType: "user_group",
-					StreamID:   streamID,
-					EventType:  string(eventtypes.UserGroupMemberRemoved),
-					Data: map[string]any{
-						"group_id": groupID,
-						"user_id":  userID,
-					},
-					ActorType: "scim",
-					ActorID:   provider.ID,
-				})
-			}
-		}
-	}
-}
-
 // deleteGroup handles DELETE /scim/v2/{slug}/Groups/{id}
 func (h *Handler) deleteGroup(w http.ResponseWriter, r *http.Request) {
 	h.logger.Debug("SCIM deleteGroup called")
