+++ internal/scim/groups_create.go
@@ -1,0 +1,219 @@
+// SCIM createGroup handler. Extracted from groups.go (audit F009 /
+// #149, slice 5) so groups.go stays under the issue's <500 LOC bar.
+//
+// createGroup is the most-branched group handler — handles SCIM
+// clients that re-POST on every sync cycle by syncing display name
+// + members on idempotent re-POST instead of returning 409.
+package scim
+
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+	"net/http"
+
+	"github.com/jackc/pgx/v5"
+
+	"github.com/manchtools/power-manage/server/internal/eventtypes"
+	"github.com/manchtools/power-manage/server/internal/store"
+	db "github.com/manchtools/power-manage/server/internal/store/generated"
+)
+
+// createGroup handles POST /scim/v2/{slug}/Groups
+func (h *Handler) createGroup(w http.ResponseWriter, r *http.Request) {
+	h.logger.Debug("SCIM createGroup called")
+	provider, ok := providerFromContext(r.Context())
+	if !ok {
+		writeError(w, http.StatusUnauthorized, "not authenticated")
+		return
+	}
+
+	var scimGroup SCIMGroup
+	limitBody(r)
+	if err := json.NewDecoder(r.Body).Decode(&scimGroup); err != nil {
+		writeError(w, http.StatusBadRequest, "invalid JSON body")
+		return
+	}
+
+	if scimGroup.DisplayName == "" {
+		writeError(w, http.StatusBadRequest, "displayName is required")
+		return
+	}
+
+	ctx := r.Context()
+	baseURL := baseURLFromRequest(r, provider.Slug)
+
+	scimGroupID := scimGroup.ExternalID
+	if scimGroupID == "" {
+		// Use the SCIM-generated ID as the SCIM group ID if no external ID
+		scimGroupID = scimGroup.ID
+	}
+	if scimGroupID == "" {
+		scimGroupID = newULID()
+	}
+
+	// Check if this SCIM group is already mapped
+	existing, err := h.store.Queries().GetSCIMGroupMapping(ctx, db.GetSCIMGroupMappingParams{
+		ProviderID:  provider.ID,
+		ScimGroupID: scimGroupID,
+	})
+	if err == nil {
+		// Already exists — update display name if changed and return existing resource.
+		h.logger.Debug("SCIM createGroup: group already exists, syncing", "scim_group_id", scimGroupID, "user_group_id", existing.UserGroupID)
+		// This makes POST idempotent, which handles SCIM clients that re-POST on every sync.
+		if existing.ScimDisplayName != scimGroup.DisplayName {
+			h.appendEvent(ctx, store.Event{
+				StreamType: "scim_group_mapping",
+				StreamID:   existing.ID,
+				EventType:  string(eventtypes.SCIMGroupMappingUpdated),
+				Data: map[string]any{
+					"provider_id":       provider.ID,
+					"scim_group_id":     scimGroupID,
+					"scim_display_name": scimGroup.DisplayName,
+				},
+				ActorType: "scim",
+				ActorID:   provider.ID,
+			})
+			h.appendEvent(ctx, store.Event{
+				StreamType: "user_group",
+				StreamID:   existing.UserGroupID,
+				EventType:  string(eventtypes.UserGroupUpdated),
+				Data: map[string]any{
+					"name": scimGroup.DisplayName,
+				},
+				ActorType: "scim",
+				ActorID:   provider.ID,
+			})
+		}
+
+		// Reconcile members if the field was present in the JSON body.
+		// nil means the field was omitted (don't touch members).
+		// Empty slice means explicitly empty (remove all members).
+		if scimGroup.Members != nil {
+			h.reconcileGroupMembers(ctx, provider, existing.UserGroupID, scimGroup.Members)
+		}
+
+		group, err := h.buildGroupResource(ctx, provider.ID, existing, baseURL)
+		if err != nil {
+			// User group referenced by mapping no longer exists.
+			// Remove the orphaned mapping and fall through to create
+			// a fresh group+mapping pair below.
+			h.logger.Warn("orphaned SCIM group mapping, removing and recreating",
+				"mapping_id", existing.ID, "user_group_id", existing.UserGroupID, "error", err)
+			h.appendEvent(ctx, store.Event{
+				StreamType: "scim_group_mapping",
+				StreamID:   existing.ID,
+				EventType:  string(eventtypes.SCIMGroupUnmapped),
+				Data: map[string]any{
+					"provider_id":   provider.ID,
+					"scim_group_id": existing.ScimGroupID,
+				},
+				ActorType: "scim",
+				ActorID:   provider.ID,
+			})
+			// Fall through to create a new group below
+		} else {
+			writeJSON(w, http.StatusOK, group)
+			return
+		}
+	}
+	if !errors.Is(err, pgx.ErrNoRows) && err != nil {
+		h.logger.Error("failed to check existing SCIM group mapping", "error", err)
+		writeError(w, http.StatusInternalServerError, "internal server error")
+		return
+	}
+
+	// Create the user group
+	h.logger.Debug("SCIM createGroup: creating new group", "scim_group_id", scimGroupID, "display_name", scimGroup.DisplayName)
+	userGroupID := newULID()
+	err = h.store.AppendEvent(ctx, store.Event{
+		StreamType: "user_group",
+		StreamID:   userGroupID,
+		EventType:  string(eventtypes.UserGroupCreated),
+		Data: map[string]any{
+			"name":        scimGroup.DisplayName,
+			"description": fmt.Sprintf("SCIM-provisioned group from %s", provider.Name),
+		},
+		ActorType: "scim",
+		ActorID:   provider.ID,
+	})
+	if err != nil {
+		h.logger.Error("failed to create user group via SCIM", "error", err)
+		writeError(w, http.StatusInternalServerError, "failed to create group")
+		return
+	}
+
+	// Create the SCIM group mapping
+	mappingID := newULID()
+	err = h.store.AppendEvent(ctx, store.Event{
+		StreamType: "scim_group_mapping",
+		StreamID:   mappingID,
+		EventType:  string(eventtypes.SCIMGroupMapped),
+		Data: map[string]any{
+			"provider_id":       provider.ID,
+			"scim_group_id":     scimGroupID,
+			"scim_display_name": scimGroup.DisplayName,
+			"user_group_id":     userGroupID,
+		},
+		ActorType: "scim",
+		ActorID:   provider.ID,
+	})
+	if err != nil {
+		h.logger.Error("failed to create SCIM group mapping", "error", err)
+		writeError(w, http.StatusInternalServerError, "failed to map SCIM group")
+		return
+	}
+
+	// Add members if provided
+	for _, member := range scimGroup.Members {
+		if member.Value == "" {
+			continue
+		}
+		streamID := userGroupID + ":" + member.Value
+		h.appendEvent(ctx, store.Event{
+			StreamType: "user_group",
+			StreamID:   streamID,
+			EventType:  string(eventtypes.UserGroupMemberAdded),
+			Data: map[string]any{
+				"group_id": userGroupID,
+				"user_id":  member.Value,
+			},
+			ActorType: "scim",
+			ActorID:   provider.ID,
+		})
+	}
+
+	// Build response
+	mapping, err := h.store.Queries().GetSCIMGroupMapping(ctx, db.GetSCIMGroupMappingParams{
+		ProviderID:  provider.ID,
+		ScimGroupID: scimGroupID,
+	})
+	if err != nil {
+		// Fall back to a minimal response
+		writeJSON(w, http.StatusCreated, SCIMGroup{
+			Schemas:     []string{GroupSchema},
+			ID:          userGroupID,
+			ExternalID:  scimGroupID,
+			DisplayName: scimGroup.DisplayName,
+			Meta: &SCIMMeta{
+				ResourceType: "Group",
+				Location:     baseURL + "/Groups/" + userGroupID,
+			},
+		})
+		return
+	}
+
+	group, err := h.buildGroupResource(ctx, provider.ID, mapping, baseURL)
+	if err != nil {
+		h.logger.Error("failed to build group resource after create", "error", err)
+		writeJSON(w, http.StatusCreated, SCIMGroup{
+			Schemas:     []string{GroupSchema},
+			ID:          userGroupID,
+			ExternalID:  scimGroupID,
+			DisplayName: scimGroup.DisplayName,
+		})
+		return
+	}
+
+	writeJSON(w, http.StatusCreated, group)
+}
